more spoj codes added

+++ SPOJ/TRICOUNT/count_tri.cpp
@@ -1,12 +1,14 @@
 #include <iostream>
+#include <cmath>
+#include <climits>
 int main(){
-    int N, n;
-    long long unsigned int ans;
+    long long int N, n;
+    unsigned long long int ans;
     
     std::cin >> N;
-    for(int i=0; i<N; i++){
+    for(long long int i=0; i<N; i++){
        std::cin >> n;
-       ans = n*(n+2)*(2*n+1)/8;
+       ans = n*(n+2)*(2*n+1)/8;      
        std::cout << ans << std::endl;
     }
 }
